+++ CarRental.Domain/Models/Vehicle.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; init; }
         public string Model { get; set; }
-        public int PricePerDay { get; set; }
+        public decimal PricePerDay { get; set; }
         public bool Active { get; set; }
     }
 }
