+++ CarRental.WebAPI/DTOs/Vehicle/GetVehicleResponse.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Model { get; set; }
-        public int PricePerDay { get; set; }
+        public decimal PricePerDay { get; set; }
         public bool Active { get; set; }
 
         public GetVehicleResponse FromDomain(Domain.Models.Vehicle vehicle)
